+++ src/calendar/demos/mobile.vue
@@ -1,9 +1,9 @@
 <template>
   <div class="tdesign-mobile-demo">
     <h1 class="title">Calendar 日历</h1>
-    <p class="summary">按钮用于选择日期。</p>
+    <p class="summary">按照日历形式展示数据或日期的容器</p>
 
-    <tdesign-demo-block title="01 基础" summary="基础用法">
+    <tdesign-demo-block title="01 组件类型">
       <div class="Calendar-demo">
         <BaseCalendarDemo />
         <multipleVue />
@@ -11,29 +11,15 @@
       </div>
     </tdesign-demo-block>
 
-    <tdesign-demo-block title="02 自定义" summary=" ">
+    <tdesign-demo-block title="02 组件样式" summary="可以自由定义想要的风格">
       <customTextDemo />
       <customButtonVue />
       <customRangeVue />
     </tdesign-demo-block>
 
-    <!-- <tdesign-demo-block title="03 设置标题" summary="自定义标题">
-      <div class="Calendar-demo">
-        <HeadCalendarDemo />
-      </div>
+    <tdesign-demo-block summary="不使用 Popup">
+      <withoutPopup />
     </tdesign-demo-block>
-
-    <tdesign-demo-block title="04 设置单元格" summary="自定义单元格">
-      <div class="Calendar-demo">
-        <CellCalendarDemo />
-      </div>
-    </tdesign-demo-block>
-
-    <tdesign-demo-block title="04 设置背景颜色" summary="自定义单元格和按钮颜色">
-      <div class="Calendar-demo">
-        <ColorCalendarDemo />
-      </div>
-    </tdesign-demo-block> -->
   </div>
 </template>
 <script lang="ts" setup>
@@ -43,4 +29,5 @@ import customTextDemo from './custom-text.vue';
 import rangeDemo from './range.vue';
 import customButtonVue from './custom-button.vue';
 import customRangeVue from './custom-range.vue';
+import withoutPopup from './without-popup.vue';
 </script>
