+++ src/calendar/demos/without-popup.vue
@@ -1,0 +1,9 @@
+<template>
+  <t-calendar :use-popup="false" type="multiple" @select="handleSelect" />
+</template>
+
+<script lang="ts" setup>
+const handleSelect = (val: any) => {
+  console.log(val);
+};
+</script>
